feat: extract trust registry
